Scenario 2 complete

+++ www/js/services.js
@@ -1,8 +1,8 @@
 bddApp.service('calculatorService', function(){
   var add = function(operand1, operand2) {
 	if((operand1 != undefined)&&(operand2 != undefined)){
-		if(parseInt(operand1) && parseInt(operand2)){
-			return parseInt(operand1) + parseInt(operand2);
+		if(parseFloat(operand1) && parseFloat(operand2)){
+			return parseFloat(operand1) + parseFloat(operand2);
 		}else{
 			return 0;
 		}
